% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstFunctionRcs.R
\name{CleaningTransformingDf}
\alias{CleaningTransformingDf}
\title{Preparación de datos para hacerlos datos geoespaciales}
\usage{
CleaningTransformingDf(DataFrame, SpatialReference, latitude = 0, variables)
}
\arguments{
\item{DataFrame}{Conjunto de datos a procesar.}

\item{SpatialReference}{Referencia espacial a asignar.}

\item{latitude}{Punto a partir del cual se considerará la latitud, por default es 0.}

\item{variables}{Variables que son de interes, en forma de vector, con entradas de tipo string.}
}
\value{
Conjunto de datos listos para usarse como datos geoespaciales
}
\description{
Preparación de datos para hacerlos datos geoespaciales
}
