% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_SimulacionPseudoausenciasFunction.R
\name{SimulatePseudoabsencesNoNAListDf}
\alias{SimulatePseudoabsencesNoNAListDf}
\title{Generación de Pseudo-ausencias para Múltiples Especies}
\usage{
SimulatePseudoabsencesNoNAListDf(
  Resultados,
  RasterCrsReferencia,
  ObjetoCrsReferencia,
  RadioGradosExclusion = 0.001,
  RadioMetrosExclusion = 100,
  NombreVariablesAmbientales,
  verbose = TRUE,
  seed = NULL
)
}
\arguments{
\item{Resultados}{Lista de data.frames con los datos de presencia por especie.}

\item{RasterCrsReferencia}{Objeto \code{SpatRaster} que define la extensión y resolución.}

\item{ObjetoCrsReferencia}{Sistema de referencia de coordenadas (CRS) de destino.}

\item{RadioGradosExclusion}{Radio de exclusión si el CRS es geográfico. Por defecto \code{0.001}.}

\item{RadioMetrosExclusion}{Radio de exclusión si el CRS es proyectado. Por defecto \code{100}.}

\item{NombreVariablesAmbientales}{Vector con los nombres de los predictores.}

\item{verbose}{Logico. Muestra el progreso en consola.}

\item{seed}{Semilla para reproducibilidad.}
}
\value{
Una lista de objetos \code{sf} con datos de presencia (1) y pseudo-ausencia (0).
}
\description{
Genera puntos de pseudo-ausencia para una lista de especies, evitando áreas
cercanas a las presencias conocidas mediante un radio de exclusión.
}
