% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningTransformingDfToSf_function.R
\name{CleaningTransformingDfToSf}
\alias{CleaningTransformingDfToSf}
\title{Preparación de datos para hacerlos geoespaciales
Preparation of data to transform its in geospacial data}
\usage{
CleaningTransformingDfToSf(
  DataFrame,
  SpatialReference,
  latitude = 0,
  variables
)
}
\arguments{
\item{DataFrame}{Conjunto de datos a procesa / Data set.}

\item{SpatialReference}{Referencia espacial a asignar / Spatial references to obtain the reference CRS.}

\item{latitude}{Punto a partir del cual se considerará la latitud, por default es 0 / Point from which latitude is considered, by default it is 0.}

\item{variables}{Variables que son de interes, en forma de vector, con entradas de tipo string / Dependent variables.}
}
\value{
Un objeto \code{sf} (Simple Features) listo para análisis geoespacial que incluye:
\itemize{
\item Las variables seleccionadas
\item Geometrías espaciales transformadas
\item Una nueva columna \code{espcvld} que cuenta las observaciones por especie
}
}
\description{
Preparación de datos para hacerlos geoespaciales
Preparation of data to transform its in geospacial data
}
\details{
La función realiza los siguientes pasos:
\enumerate{
\item Filtra observaciones al norte de la latitud especificada
\item Convierte el DataFrame a objeto \code{sf} con coordenadas geográficas (CRS: 4326)
\item Transforma las coordenadas al sistema de referencia espacial especificado
\item Recorta los datos usando la intersección con la referencia espacial
\item Elimina duplicados espaciales basados en coordenadas y especie
\item Agrega una columna con el conteo de observaciones por especie (espcvld)
}
}
\examples{
\dontrun{
# Crear datos de ejemplo
set.seed(123)
datos_ejemplo <- data.frame(
  longitud = runif(100, -100, -80),
  latitud = runif(100, 15, 30),
  especie = sample(c("Quercus", "Pinus", "Abies"), 100, replace = TRUE),
  altura = runif(100, 5, 25),
  diametro = runif(100, 10, 50)
)

# Crear objeto espacial de referencia
library(sf)
referencia_espacial <- st_as_sfc(st_bbox(c(
  xmin = -95, xmax = -85,
  ymin = 20, ymax = 25
), crs = 4326))
referencia_espacial <- st_transform(referencia_espacial, 32614)

# Definir variables de interés
vars_interes <- c("especie", "altura", "diametro")

# Usar la función
datos_procesados <- CleaningTransformingDfToSf(
  DataFrame = datos_ejemplo,
  SpatialReference = referencia_espacial,
  latitude = 18,
  variables = vars_interes
)

# Ver resultados
print(datos_procesados)
summary(datos_procesados$espcvld)
}

}
